-add auto size up

+++ lib/src/main/java/cyoap_main/design/platform/AbstractPlatform.java
@@ -28,10 +28,10 @@ public class AbstractPlatform {
 
     public double local_x = 0;// local position of screen
     public double local_y = 0;
-    public int min_x = -800;
-    public int min_y = -1600;
-    public int max_x = 800;
-    public int max_y = 1600;
+    public float min_x = -800;
+    public float min_y = -1600;
+    public float max_x = 800;
+    public float max_y = 1600;
     public float scale = 1.0f;
     @JsonIgnore
     public float maximize = 2f;
